Fix issue #50
When a statement contains several queries separated by a semicolon character
and at least one of these queries contains a string value including a semicolon
the list of queries to execute was incorrect resulting into a syntax error
during the execution of the split statements.

The statements splitting has been refactored and is applied in any compliance
mode.